chunk_pipeline: job split logic
Previous approach sloppily used the ceiling function, which
overallocated work to some jobs, and resulted in a small percent of jobs
receiving no work.

For instance a N_periods=70k grid might have a few percent overhead
from this choice.

This commit fixes this by allocating `base+1` periods to a subset of
chunks, and `base` periods to the rest.