add heuristics for periodogram postprocessing
The general problem of separating the singular sharp planet peak from
the forest of broad rotation peaks is, admittedly, outstanding.

The current approach is a mix of weird heuristics.

No matter what, the current default is to use a mean slider to detrend.

In addition, if the star rotates "fast" (Prot<3 days), then the peaks
become narrow, and a variable length sliding window (mean) detrending is
applied.

However, the higher order harmonics (like >10x harmonics) of the fast
rotators are still present, and when you go higher and higher those
harmonics deviate further from the "fundmanetal".  In addition,
differential rotation and spot evolution can yield __multi-peaked__
broad rotation signals, as in the case of KOI-7368 / Kepler-1974.

A wavelet approach for measuring "Prot" if you want to use it in
smoothing would probably be a better idea than the current default
of Lomb Scargle.

However, more broadly, this seems like a classification problem.
Maybe exploring classifiers would be a good idea.